fix: missed refactoring
