+++ pkg/helpers/cni_test.go
@@ -3,7 +3,7 @@ package helpers
 import (
 	"testing"
 
-	"github.com/kubesimplify/ksctl/pkg/utils/consts"
+	"github.com/kubesimplify/ksctl/pkg/helpers/consts"
 )
 
 func FuzzValidateCNI(f *testing.F) {
